Move new errors from "playing"
moves the new APIException structures from playing
package into the normal exceptions package
