+++ src/main/java/io/stargate/sgv2/jsonapi/exception/DocumentException.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 /**
  * Errors related to the document user is trying to insert or update, could be for collections or
