+++ src/main/java/io/stargate/sgv2/jsonapi/exception/EmbeddingProviderException.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 public class EmbeddingProviderException extends ServerException {
