+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 import com.google.common.base.Preconditions;
 import java.util.HashMap;
