+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ErrorFamily.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 /**
  * Top level hierarchy for all exceptions thrown by the API.
