+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ErrorInstance.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 import java.util.Optional;
 import java.util.UUID;
