+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ErrorScope.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 /**
  * Interface for any enum that represents an error scope to implement.
