+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ErrorTemplate.java
@@ -1,6 +1,5 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
-import io.stargate.sgv2.jsonapi.exception.ErrorCodeV1;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
