+++ src/main/java/io/stargate/sgv2/jsonapi/exception/ExceptionHandler.java
