+++ src/main/java/io/stargate/sgv2/jsonapi/exception/FilterException.java
@@ -1,4 +1,4 @@
-package io.stargate.sgv2.jsonapi.exception.playing;
+package io.stargate.sgv2.jsonapi.exception;
 
 /**
  * Errors related to the filter clause in a request.
